requirements.txt
streamlit
torch
torchvision
Pillow
pandas
matplotlib
numpy